move auth to modules
