+++ server/users/src/modules/auth/auth.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 import { PassportModule } from '@nestjs/passport';
 import { AuthService } from './services/auth.service';
-import { UsersModule } from '../modules/users/users.module';
-import { UtilsModule } from './../utils/utils.module';
+import { UsersModule } from '../users/users.module';
+import { UtilsModule } from '../../utils/utils.module';
 import { LocalStrategy } from './strategies/local.strategy';
 import { AuthController } from './controllers/auth.controllers';
 import { JwtModule } from '@nestjs/jwt';
