+++ server/users/src/modules/auth/guards/jwt-auth.guard.ts
